Clear cache while adding a new feed
Before, the feed cache was preventing to add a feed when using twice the
bookmarklet if the first time the user aborted the process.
Now, the cache is removed for the feed before loading information. This
behavior as a limited impact since only the selected feed cache is removed.

See #3658