Fix Docker documents for port mapping and caddy use.
Apache has a well known limitation when used in a docker container
that the published port must match the listening port. If this isn't
the case, content will fail to serve correctly and pages come up blank
instead.

Update the example docker compose files and example code to
identify that the PUBLISH_PORT needs to match the LISTEN port.

Add a Caddy reverse proxy example matching the Nginx one.