Remove new window target in normal view
As of v1.24.3, clicking the title of an article in normal view opens a
new tab in the browser. Remove the link reference to expand the article
within the reader, while still allowing for direct external view via
the "See on original website" link at the end of the row.