Dockerfile: switch from CGI to FastCGI for PHP
This installs and runs PHP-FPM (FastCGI process manager).

Both Apache2 and PHP-FPM are put into background as a shell job,
such that "wait -n" can be used to terminate if either of both
exits.

If any of the jobs outputs to standard out/err, it will go to
the Docker log.